AppPackage & NavxManifest
