+++ extension/package.json
@@ -495,7 +495,8 @@
         "minimatch": "^3.0.4",
         "node-powershell": "^3.3.1",
         "time-stamp": "^1.1.0",
-        "xmldom": "^0.5.0"
+        "xmldom": "^0.5.0",
+        "txml": "^4.0.1"
     },
     "devDependencies": {
         "@types/jdataview": "0.0.31",
