+++ extension/src/SymbolReference/Interfaces/AppPackage.ts
@@ -1,0 +1,9 @@
+import { ALObject } from "../../ALObject/ALObject";
+import { ManifestPackage } from "./NavxManifest";
+import { SymbolReference } from "./SymbolReference";
+
+export interface AppPackage {
+    manifest: ManifestPackage;
+    symbolReference?: SymbolReference;
+    objects?: ALObject[];
+}
