+++ extension/src/test/Symbols.test.ts
@@ -1,16 +1,33 @@
+import * as path from 'path';
 import * as assert from 'assert';
 import * as SymbolReferenceReader from '../SymbolReference/SymbolReferenceReader';
 
+const testResourcesPath = '../../src/test/resources/.alpackages';
+
+let baseAppPath = path.resolve(__dirname, testResourcesPath, 'Microsoft_Base Application_18.0.23013.23320.app');
+let testAppPath = path.resolve(__dirname, testResourcesPath, 'Default publisher_Al_1.0.0.0.app');
+
+
 suite("Symbol Parsing", function () {
-    test.only("Spider", function () {
-        const appFilePath = 'D:\\VSCode\\Git\\GitHub\\nab-al-tools\\test-app\\Xliff-test\\Default publisher_Al_1.0.0.0.app';
-        const testAppSymbols = JSON.parse(SymbolReferenceReader.getSymbolReferenceFromAppFile(appFilePath))
-        // const symbolData = JSON.parse(SymbolReader.getSymbolReferenceFromAppFile('D:\\VSCode\\Git\\NAB\\Spider\\App\\.alpackages\\SmartApps_Spider_18.0.21356.0.app'));
-        // assert.equal(symbolData.Name, 'Spider');
-        assert.equal(testAppSymbols.Name, 'Al');
-        const objects = SymbolReferenceReader.getObjectsFromAppFile(appFilePath);
-        assert.equal(objects.length, 2, 'unexpected number of objects');
-        assert.equal(objects[0].name, 'NAB Test Table', 'unexpected name');
+    test.only("TestApp", function () {
+        const appPackage = SymbolReferenceReader.getObjectsFromAppFile(testAppPath);
+        assert.equal(appPackage.manifest.App[0]._attributes.Name, 'Al');
+        if (appPackage.objects) {
+            assert.equal(appPackage.objects.length, 2, 'unexpected number of objects');
+            assert.equal(appPackage.objects[0].name, 'NAB Test Table', 'unexpected table name');
+        } else {
+            assert.fail('No objects found')
+        }
+    });
+    test.only("BaseApp", function () {
+        const appPackage = SymbolReferenceReader.getObjectsFromAppFile(baseAppPath);
+        assert.equal(appPackage.manifest.App[0]._attributes.Name, 'Base Application');
+        if (appPackage.objects) {
+            assert.equal(appPackage.objects.length, 1468, 'unexpected number of objects');
+            assert.equal(appPackage.objects[0].name, 'AAD Application', 'unexpected table name');
+        } else {
+            assert.fail('No objects found')
+        }
     });
 
 });
